Moved theme package from ui package to notesapplication package
