+++ app/src/main/java/com/atakanmadanoglu/notesapplication/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.atakanmadanoglu.notesapplication.ui.theme
+package com.atakanmadanoglu.notesapplication.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
